requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0
st-login-form
st-supabase-connection